requirements.txt
flask
pdf2image
flask_wtf
opencv-python
spacy
nltk
pymystem3
flask-cors
pillow
psutil
gunicorn
pytesseract
pymongo
